ref(tools): Move zod tool adapter to support
Keep the Zod-backed tool authoring adapter with the other tool-support code instead of the tool definition module. Update internal tool imports and move the focused unit test to match the new ownership.
