+++ packages/junior/src/chat/slack/tools/list/add-items.ts
@@ -4,7 +4,7 @@ import {
   slackUserIdParam,
 } from "@/chat/slack/id-param";
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 import { ToolInputError } from "@/chat/tools/execution/tool-input-error";
 import { createOperationKey } from "@/chat/tools/idempotency";
 import type { ToolState } from "@/chat/tools/types";
