+++ packages/junior/src/chat/slack/tools/list/create.ts
@@ -1,6 +1,6 @@
 import { createTodoList } from "@/chat/slack/tools/list/api";
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 import { createOperationKey } from "@/chat/tools/idempotency";
 import type { ToolState } from "@/chat/tools/types";
