+++ packages/junior/src/chat/slack/tools/list/get-items.ts
@@ -1,6 +1,6 @@
 import { listItems } from "@/chat/slack/tools/list/api";
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 import type { ToolState } from "@/chat/tools/types";
 
 /** Create a tool that reads items from the active Slack list. */
