+++ packages/junior/src/chat/tools/sandbox/edit-file.ts
@@ -15,7 +15,7 @@ import {
   type TextReplacement,
 } from "@/chat/tools/sandbox/text-edits";
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 
 type EditReplacement = TextReplacement;
