+++ packages/junior/src/chat/tools/sandbox/read-file.ts
@@ -3,7 +3,7 @@ import {
   positiveInteger,
 } from "@/chat/tools/sandbox/file-utils";
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 
 const DEFAULT_READ_LIMIT = 1000;
