+++ packages/junior/src/chat/tools/system-time.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { zodTool } from "@/chat/tools/definition";
+import { zodTool } from "@/chat/tool-support/zod-tool";
 
 export function createSystemTimeTool() {
   return zodTool({
